requirements.txt
Flask==2.3.4
Werkzeug==2.3.6
openpyxl==3.1.3
Pillow==10.1.0